% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreetMaps.R
\name{Maps_api}
\alias{Maps_api}
\title{Draw the municipality boundaries}
\usage{
Maps_api(dataset)
}
\arguments{
\item{dataset}{one of the following (fi-8, ch-8, no-4, no-7, dk-4, se-4, se-7, us-4, gl-7,world-2)}
}
\value{
The object of class Maps_api.
}
\description{
Draw the municipality boundaries
}
\examples{
\dontrun{
Maps_api("se-4")
Maps_api("se-7")}
}
